feat: add funcs to get stations that are together
Since WMATA codes each station with multiple lines that do not share tracks as separate stations in their API, provide methods to get stations at the same location in an array.

This provides to functions:
- `together()` which gets the other stations at the same location
- `allTogether()` which appends the current station to the results of `together()`

The use case for these methods is to to get all nextTrain results for a single multi-platform station (Fort Totten, Gallery Pl-Chinatown, L'Enfant Plaze, or Metro Center) in a single API call.